[Refactor] - annotated responsibilities
Here we see that even a seemingly small code block can contain a number of
responsibility smells. The Game object handles user input, manages the state of
the game (right now just the cascades), and checks for the legality of moves.
This leads to methods that have several different responsibilities at once.